+++ NEWS.md
@@ -1,7 +1,11 @@
+# territorial 0.8.12 (2026/08/12)
+- Nueva función `contar_comunas()` para revisar rápidamente si faltan comunas en una tabla de datos o vector, o confirmar que estén todas.
+- Menos mensajes irrelevantes en `validar_comunas()`.
+- Se eliminan `poblacion_proyeccion` y `agregar_poblacion()` porque se traspasan al [nuevo paquete `{poblador}` especializado en datos de población de Chile.](https://bastianolea.github.io/poblador/)
+
 # territorial 0.8.11 (2026/08/11)
 - Nueva función `buscar_comuna()`, que facilita la búsqueda de comunas por texto parcial o inexacto en cualquier tabal de datos, o por defecto en `territorial::territorios`.
 - Agregados ejemplos a `as_nombre_comuna()` y `as_nombre_region()`, gracias a `pkgcheck::pkgcheck()`
-- Agregada participación de ROpenSci a DESCRIPTION
 
 # territorial 0.8.9.2 (2026/08/06)
 - Corrección de función `validar_comunas()` para arrojar error en caso de recibir listas